requirements.txt
# --- MOTOR PRINCIPAL ---
pandas>=2.2.0
numpy>=1.26.0
XlsxWriter>=3.2.0
python-calamine>=0.2.0
openpyxl>=3.1.2
PyYAML>=6.0.1
python-dotenv==1.0.0

# --- QUALIDADE E TESTES (NOVOS) ---
pytest>=8.0.0        # Para rodar os testes
pandera>=0.18.0      # Para validar colunas do Excel